Add Grid support.
Notably Grid doesn't use FileType stuff. I played with
requesting a column. But the meta data feels quite twisted.
